requirements.txt
python-binance>=1.0.19
websockets>=11.0.3